設定ページを階層構造にリファクタリング
カスタマイズ（通知・パーソナライズ）とアカウント（プラン・サブスク管理・削除）を
サブページに分離し、メインページからナビゲーションで遷移する構成に変更。
